Miecz Re-Fluff
I'm not sure why, but the Miecz is referred to as a "Support Weapon", whatever that is in this context, while its own magazine claims to belong to a Submachinegun.

This just alters some flavourtext to make the entire thing coherent again, I've made no changes to the actual weapon itself.